refactor(Dotenv): put Dotenv related objects into helpers.dotenv
