+++ src/main/scala/helper/dotenv/Dotenv.scala
@@ -1,12 +1,11 @@
-package helper
+package helper.dotenv
 
 import scala.io.Source
-import scala.collection.JavaConverters._
 import scala.util.{Failure, Success, Try}
 
 object Dotenv {
   // To load .env file to the Environmental Variables for development
-  def loadEnv (envFilePath: String = ".env"): Unit = {
+  def loadEnv(envFilePath: String = ".env"): Unit = {
     var envLines: List[String] = Try(Source.fromResource(envFilePath).getLines().toList) match {
       case Success(lines) => lines
       case Failure(e) => {
@@ -34,4 +33,3 @@ object Dotenv {
     }
   }
 }
-
